Send event on account unlink (#2379)
* Send event on account unlink

* Move unlink to its own module
