+++ init_app.py
@@ -6,7 +6,7 @@ from app.db import Session
 from app.log import LOG
 from app.models import Mailbox, Contact, SLDomain, Partner
 from app.pgp_utils import load_public_key
-from app.proton.utils import PROTON_PARTNER_NAME
+from app.proton.proton_partner import PROTON_PARTNER_NAME
 from server import create_light_app
