+++ tests/cron/test_get_alias_for_hibp.py
@@ -14,7 +14,7 @@ from app.models import (
     PartnerSubscription,
     User,
 )
-from app.proton.utils import get_proton_partner
+from app.proton.proton_partner import get_proton_partner
 from tests.utils import create_new_user, random_token
