+++ tests/test_alias_suffixes.py
@@ -3,7 +3,7 @@ import re
 from app.alias_suffix import get_alias_suffixes
 from app.db import Session
 from app.models import SLDomain, PartnerUser, AliasOptions, CustomDomain
-from app.proton.utils import get_proton_partner
+from app.proton.proton_partner import get_proton_partner
 from init_app import add_sl_domains
 from tests.utils import create_new_user, random_token
